* Fix unit tests
