+++ api/src/test/java/org/openmrs/module/disa/api/sync/FinalLabResultHandlerTest.java
@@ -3,7 +3,6 @@ package org.openmrs.module.disa.api.sync;
 import static org.hamcrest.Matchers.anEmptyMap;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
-import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -18,7 +17,6 @@ import org.openmrs.Location;
 import org.openmrs.Patient;
 import org.openmrs.api.EncounterService;
 import org.openmrs.api.LocationService;
-import org.openmrs.module.disa.FsrLog;
 import org.openmrs.module.disa.HIVVLLabResult;
 import org.openmrs.module.disa.LabResult;
 import org.openmrs.module.disa.LabResultStatus;
@@ -76,72 +74,6 @@ public class FinalLabResultHandlerTest extends BaseContextMockTest {
         finalLabResultHandler.handle(labResult);
 
         verify(next, never()).handle(labResult);
-    }
-
-    @Test
-    public void shouldSaveTheEncounterForProcessedLabResults() {
-        Encounter encounter = new Encounter();
-        Patient patient = new Patient(1);
-        encounter.setPatient(patient);
-        finalLabResultHandler.getSyncContext().put(BaseLabResultHandler.ENCOUNTER_KEY, encounter);
-
-        LabResult labResult = new HIVVLLabResult();
-        labResult.setRequestId("MZDISAPQM0000000");
-        labResult.setNid("000000000/0000/00000");
-        labResult.setLabResultStatus(LabResultStatus.PROCESSED);
-
-        Location location = new Location();
-        location.setUuid("93377be8-1093-47f0-ad05-e014d3a14615");
-        when(locationService.getDefaultLocation()).thenReturn(location);
-
-        finalLabResultHandler.handle(labResult);
-
-        verify(encounterService, times(1)).saveEncounter(encounter);
-    }
-
-    @Test
-    public void shouldCreateFsrLogForProcessedLabResults() {
-        Encounter encounter = new Encounter();
-        Patient patient = new Patient(1);
-        encounter.setPatient(patient);
-        finalLabResultHandler.getSyncContext().put(BaseLabResultHandler.ENCOUNTER_KEY, encounter);
-
-        LabResult labResult = new HIVVLLabResult();
-        labResult.setRequestId("MZDISAPQM0000000");
-        labResult.setNid("000000000/0000/00000");
-        labResult.setLabResultStatus(LabResultStatus.PROCESSED);
-
-        Location location = new Location();
-        location.setUuid("93377be8-1093-47f0-ad05-e014d3a14615");
-        when(locationService.getDefaultLocation()).thenReturn(location);
-
-        finalLabResultHandler.handle(labResult);
-
-        verify(disaService, times(1)).saveFsrLog(any(FsrLog.class));
-    }
-
-    @Test
-    public void shouldUpdateProcessedLabResults() {
-
-        Encounter encounter = new Encounter();
-        Patient patient = new Patient(1);
-        encounter.setPatient(patient);
-        finalLabResultHandler.getSyncContext().put(BaseLabResultHandler.ENCOUNTER_KEY, encounter);
-
-        LabResult labResult = new HIVVLLabResult();
-        labResult.setRequestId("MZDISAPQM0000000");
-        labResult.setNid("000000000/0000/00000");
-        labResult.setLabResultStatus(LabResultStatus.PROCESSED);
-
-        Location defaultLocation = new Location();
-        defaultLocation.setUuid("93377be8-1093-47f0-ad05-e014d3a14615");
-        when(locationService.getDefaultLocation()).thenReturn(defaultLocation);
-
-        finalLabResultHandler.handle(labResult);
-
-        assertThat(labResult.getLabResultStatus(), is(LabResultStatus.PROCESSED));
-        assertThat(labResult.getSynchronizedBy(), is(defaultLocation.getUuid()));
-        verify(labResultService, times(1)).updateLabResult(any(LabResult.class));
     }
 
     @Test
